requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.1
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.12.4
filelock==3.4.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.2
Markdown==3.3.6
oauthlib==3.1.1
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
urllib3==1.26.7
virtualenv==20.10.0
whitenoise==5.3.0
zipp==3.6.0
django-googledrive-storage==1.6.0
google-api-core==2.3.0
google-api-python-client==2.33.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
mysqlclient==2.1.0